LPS-88988 Using a dynamic capability to hook into DM
