+++ modules/apps/adaptive-media/adaptive-media-document-library/src/main/java/com/liferay/adaptive/media/document/library/internal/repository/capabilities/AMPortletRepositoryCapability.java
@@ -1,0 +1,29 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.adaptive.media.document.library.internal.repository.capabilities;
+
+import com.liferay.portal.kernel.repository.capabilities.Capability;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Alejandro Tardín
+ */
+@Component(
+	property = "repository.class.name=com.liferay.portal.repository.portletrepository.PortletRepository",
+	service = Capability.class
+)
+public class AMPortletRepositoryCapability extends BaseCapability {
+}
